+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ## Features
 
+* Add #PAC-227: Dynamic Handling of columns for import of EAV attributes
 * Remove stack trace of exception for missing media directories > log a simple debug message instead
 
 # Version 20.1.1